close: Via Tps helper | Drop: via Tps helper | refactor ws_client
